# apps/backend/openapi.yaml
openapi: "3.0.3"
info:
  title: BMad Todo API
  version: "0.1.0"

servers:
  - url: http://localhost:3000
    description: Local development

components:
  schemas:
    Todo:
      type: object
      required: [id, text, completed, createdAt]
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        completed:
          type: boolean
        createdAt:
          type: string
          format: date-time

    TodoResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Todo"

    TodoListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Todo"

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string

paths:
  /api/todos/{id}:
    patch:
      summary: Toggle todo completion
      operationId: toggleTodo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [completed]
              properties:
                completed:
                  type: boolean
                  example: true
      responses:
        "200":
          description: Todo updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /health:
    get:
      summary: Health check
      operationId: healthCheck
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok

  /api/todos:
    get:
      summary: List all todos
      operationId: listTodos
      responses:
        "200":
          description: Array of todos ordered by creation date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoListResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      summary: Create a todo
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  minLength: 1
                  maxLength: 500
                  pattern: "\\S"
                  example: Buy groceries
      responses:
        "201":
          description: Todo created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"